Feature/UI

+++ fuel/app/classes/model/task.php
@@ -2,16 +2,23 @@
 
 class Model_Task extends \Model
 {
-    public static function get_tasks_all()
-    {
-        return \DB::select()
-            ->from('tasks')
-            ->order_by('created_at', 'desc')
-            ->execute()
-            ->as_array();
-    }
+  public static function get_tasks_all($user_id)
+  {
+    return \db::select(
+      'tasks.*',
+      array(\db::expr('count(reference_items.id)'), 'reference_items_count')
+    )
+    ->from('tasks')
+    ->join('reference_items', 'left')
+    ->on('tasks.id', '=', 'reference_items.task_id')
+    ->where('tasks.user_id', '=', $user_id)
+    ->group_by('tasks.id')
+    ->order_by('created_at', 'desc')
+    ->execute()
+    ->as_array();
+  }
 
-    public static function get_task_by_id($task_id)
+    public static function get_task_by_id($task_id, $user_id)
     {
         return \DB::select(
             'tasks.*',
@@ -20,13 +27,20 @@ class Model_Task extends \Model
             ->from('tasks')
             ->join('users')
             ->on('tasks.user_id', '=', 'users.id')
-            ->where('tasks.id', $task_id)
+            ->where('tasks.id', '=', $task_id)
+            ->where('tasks.user_id', '=', $user_id)
             ->execute()
             ->current();
     }
 
-    public static function create_task($title, $description, $status, $dev_location, $deadline, $user_id)
-    {
+    public static function create_task(
+        $title,
+        $description,
+        $status,
+        $dev_location,
+        $deadline,
+        $user_id
+    ) {
         return \DB::insert('tasks')
             ->set(array(
                 'title' => $title,
@@ -41,9 +55,15 @@ class Model_Task extends \Model
             ->execute();
     }
 
-
-    public static function update_task($task_id, $title, $description, $status, $dev_location, $deadline)
-    {
+    public static function update_task(
+        $task_id,
+        $title,
+        $description,
+        $status,
+        $dev_location,
+        $deadline,
+        $user_id
+    ) {
         return \DB::update('tasks')
             ->set(array(
                 'title' => $title,
@@ -53,16 +73,36 @@ class Model_Task extends \Model
                 'deadline' => $deadline,
                 'updated_at' => date('Y-m-d H:i:s'),
             ))
-            ->where('id', $task_id)
+            ->where('id', '=', $task_id)
+            ->where('user_id', '=', $user_id)
             ->execute();
     }
 
-
-    public static function delete_task($task_id)
+    public static function delete_task($task_id, $user_id)
     {
         return \DB::delete('tasks')
-            ->where('id', $task_id)
+            ->where('id', '=', $task_id)
+            ->where('user_id', '=', $user_id)
             ->execute();
     }
 
+  public static function search_tasks_by_keyword($user_id, $keyword)
+  {
+    return \DB::select(
+      'tasks.*',
+      array(\DB::expr('count(reference_items.id)'), 'reference_items_count')
+    )
+    ->from('tasks')
+    ->join('reference_items', 'left') 
+    ->on('tasks.id', '=', 'reference_items.task_id')
+    ->where('tasks.user_id', '=', $user_id)
+    ->where_open()
+      ->where('tasks.title', 'like', '%' . $keyword . '%')
+      ->or_where('tasks.description', 'like', '%' . $keyword . '%')
+    ->where_close()
+    ->group_by('tasks.id')
+    ->order_by('tasks.created_at', 'desc')
+    ->execute()
+    ->as_array();
+  }
 }
